% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obo_helper.R
\name{corpus_rebuild}
\alias{corpus_rebuild}
\title{Rebuild corpus datasets.}
\usage{
corpus_rebuild(obo, reload = TRUE, ...)
}
\arguments{
\item{obo}{An OBO interface object.}

\item{...}{Further arguments passed to load_corpus. Ignored if reload is FALSE.}

\item{realod}{Logical indicating whether the new corpus files should be reloaded.}
}
\value{
If reload, the value of load_corpus. NULL otherwise.
}
\description{
This function will rebuild all corpus datasets using the parameters currently found in the given
obo object's configuration.
}
\details{
The produced data sets will be dumped into the current configuration's data directory, using the
configured filenames.
}
