% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{count_to_pmi}
\alias{count_to_pmi}
\title{Compute a (P)PMI matrix from raw frequency counts.}
\usage{
count_to_pmi(m, positive = TRUE)
}
\arguments{
\item{m}{A (typically sparse) matrix containing cooccurrence counts}

\item{positive}{A logical value indicating whether negative values should be truncated to 0 (i.e. compute PPMI instead of PMI)}
}
\value{
The PMI value for each cell is equal to log( p(i,j) / p(i)p(j) ), i.e. the log of the observed probability over the
expected probability. The PPMI truncates negative values to 0. WARNING: In order to maintain sparsity, -Inf values in the
non-truncated case are replaced by 0.
}
\description{
Compute a (P)PMI matrix from raw frequency counts.
}
