% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{counts_to_factor}
\alias{counts_to_factor}
\title{Build a factor from max/min values in a series of variables}
\usage{
counts_to_factor(d, var, min = FALSE, conf = FALSE, drop = TRUE)
}
\arguments{
\item{d}{A data frame}

\item{var}{An optional list of variables to compare}

\item{min}{If true, use minimum instead of maximum}
}
\value{
A factor of the same length as d, where the value for each observation will correspond
        to the name of the column in var that
        contains the minimum or maximum value for that row.
}
\description{
Build a factor from max/min values in a series of variables
}
\examples{
d <- as.data.frame( matrix( rnorm( 300 ), nrow = 100, ncol = 3 ) )
names( d ) <- c( 'red', 'blue', 'green' )
count_to_factor( d )

}
