% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_util.R
\name{get_cooccurrences}
\alias{get_cooccurrences}
\title{Count cooccurrences over the given document set.}
\usage{
get_cooccurrences(obo, ds = obo$defaultDocSet())
}
\arguments{
\item{obo}{An OBO interface object.}

\item{ds}{A DocSet, built from \code{\link{make_doc_set}}.}
}
\value{
A sparse matrix containing cooccurrence counts for all terms in the given field in the
given sample of documents.
}
\description{
Produces a sparse matrix containing cooccurrence counts for all terms in the analysis field over
all documents in the given document set.
}
