% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_util.R
\name{get_index_fields}
\alias{get_index_fields}
\title{Get all fields contained in the index.}
\usage{
get_index_fields(obo)
}
\arguments{
\item{obo}{An OBO interface object.}
}
\value{
A character vector with the names of all fields present in the index.
}
\description{
Obtains a character vector with all fields contained in the corpus index.
}
\details{
The returned values may be used to e.g. list all terms in any field, construct document sets
over any one of the field's terms or combinations thereof, etc.
}
