% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_connected}
\alias{graph_connected}
\title{Determine graph connectivity up to the given tolerance}
\usage{
graph_connected(g, tol = 1)
}
\arguments{
\item{g}{An igraph graph.}

\item{tol}{An integer indicating the maximum size of detached components. Defaults to 1.}
}
\value{
\code{TRUE} if the given graph has no detached component larger than tol, \code{FALSE}
        otherwise.
}
\description{
This function will return true if the maximum component size of all components detached from the
largest component is greater or equal to the given tolerance. E.g. a tolerance of one implies
that a graph will still be considered connected if the minor components are only orphan nodes.
}
