% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_prune_connected}
\alias{graph_prune_connected}
\title{Prune a graph maintaining connectivity}
\usage{
graph_prune_connected(g, edges, tol = 1, dropV = TRUE, attr = "weight",
  desc = TRUE)
}
\arguments{
\item{g}{An igraph graph.}

\item{edges}{A vector of edges, sorted according to their significance.}

\item{tol}{An integer indicating the maximum disconnected component allowed.}

\item{dropV}{A logical indicating whether vertices in components below the given tolerance
should be dropped from the graph. Default TRUE.}

\item{attr}{An edge attribute name, used to sort edges if no edge list provided. Ignored if
edges is not NULL. Deafult 'weight' if it is.}

\item{desc}{Logical, indicating whether edges should be sorted in descending order according to
attr. Ignored if edges is not NULL, default TRUE if it is.}
}
\value{
A subgraph of g with the least significant edges removed.
}
\description{
Removes edges from a graph such that the resulting graph maintains connectivity up to the given
tolerance.
}
\details{
The given tolerance value indicates the maximum size of detached components that will be
tolerated when determining connectivity. i.e. a tolerance of one implies the graph will still be
considered connected if the size of any detached component after edge removal is no greater than
one.

Edges are removed according to the significance order provided in the given edges vector.
}
