% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obo_helper.R
\name{index_field_terms}
\alias{index_field_terms}
\title{Get all terms in the given field.}
\usage{
index_field_terms(obo, field)
}
\arguments{
\item{obo}{An OBO interface object.}

\item{field}{A field name, e.g. an element of \code{\link{get_index_fields}}.}
}
\value{
A character vector containing all the terms found in the given field.
}
\description{
Obtains a character vector with all the terms found in the given field.
}
\details{
Elements of the returned vector may be used to construct document sets over the same field
using \code{\link{make_doc_set}}.
}
