% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ints_to_str}
\alias{ints_to_str}
\title{Paste a series of integer values into a character value.}
\usage{
ints_to_str(x, flags = TRUE, sep = "")
}
\arguments{
\item{x}{An integer vector}

\item{flags}{Logical. Convert values prior to concatenation by \code{as.integer( x > 0 )}.}

\item{sep}{A character expresion tO use as separator between intger values. Empty by default.}
}
\value{
a character value containing the concatenation of x, optionally recoded as 0 and 1 and
        with the given separator in between values, if given.
}
\description{
This function takes an integer vector as input and produces a string value equal to the
concatenation of the given vector, with the given separator added in between, and optionally
interpreting the given integer values as if they were boolean flags
\code{using as.integer( x > 0 )}.
}
