% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obo_helper.R
\name{lexicon}
\alias{lexicon}
\title{Get a copy of the lexicon.}
\usage{
lexicon(obo)
}
\arguments{
\item{obo}{An OBO interface object.}
}
\value{
A lexicon data frame, with terms as row names and tf and df as columns.
}
\description{
Get a copy of the lexicon.
}
