% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_corpus}
\alias{load_corpus}
\title{Load corpus data from the given directory.}
\usage{
load_corpus(dir = getwd(), lexicon = "lexicon.dsv",
  frequencies = "frequencies.dsv", pos_counts = "pos_counts.dsv",
  cooccur = "cooccur.bin", quiet = FALSE, attach = FALSE,
  env = .GlobalEnv)
}
\arguments{
\item{dir}{Directory to find corpus data files in.}

\item{lexicon}{Lexicon file name, default 'lexicon.tsv'}

\item{frequencies}{Frequencies file name, default 'frequencies.tsv'}

\item{cooccur}{Cooccurrence matrix file name, default 'cooccur.bin'}

\item{quiet}{Logical indicating whether progress and general stats messages should be
silenced.}

\item{pos_table}{POS counts file name, default 'pos_counts.tsv'}
}
\value{
A list of length four containing the loaded corpus datasets.
}
\description{
Loads corpus data from the files found in the given directory. Corpus data includes a lexicon
file, frequency counts for all corpus partitions (if any), POS counts for every word (lemma)
in the lexicon and a cooccurrence matrix.
}
\details{
See details below for the different file formats used. DSV files are used for lexicon,
frequencies and POS counts data. Cooccurrence counts are saved as a sparse matrix stored in
triplet format as a plain array of (int, int, float) triplets. Each data file's format is
documented in its respective loading function read_*.
}
\seealso{
\code{\link{read_lexicon}}, \code{\link{read_frequencies}},
         \code{\link{read_pos_counts}} and \code{\link{read_cooccur}}.
}
