% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load_spm}
\alias{load_spm}
\title{Load sparse matrices from disk.}
\usage{
load_spm(file)
}
\arguments{
\item{file}{A character vector indicating the path to the file to be read.}
}
\value{
A Matrix::sparseMatrix built from the triplets read from the given binary file.
}
\description{
Reads a sparse matrix from binary file containing a sequence of (long, long, double) triplets.
}
\details{
This function uses a memory mapped file to read the given binary file as an array of triplets
containing the coordinates and values for non-zero entries in a sparse matrix. This format is
compatible with the binary format used natively by Glove. The component values in the triplets
are used to construct the i, j and x vectors needed by Matrix::sparseMatrix.
}
