% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_codes}
\alias{make_codes}
\title{Build factor levels from combinations of variables.}
\usage{
make_codes(d, ..., flags = TRUE, toint = TRUE, target = "code",
  mode = c("max", "comb"), drop = TRUE, ties = NA)
}
\arguments{
\item{d}{A data frame.}

\item{...}{Arguments passed to \code{\link{dplyr::select}} to choose columns to compute
levels on.}

\item{flags}{Reduce values to booleans x = x > 0. Ignored if mode is not 'comb'. This will
dramatically reduce the number of levels in the resulting factor. Default TRUE.}

\item{toint}{Convert factors to integer values. Only makes sense when flags is TRUE, in which
case the concatenation of logical values for x > 0 is interpreted as a binary
integer.}

\item{target}{Character value indicating the name of the output factor. Default: 'code'}

\item{mode}{One of 'max' or 'comb'. If 'max', value is equal to the row maximum for the
selected columns. If 'comb', value is equal to the combination of all values in
the selected column.}

\item{drop}{Logical. If TRUE, \code{\link{droplevels}} is called on the factor before return.}

\item{ties}{Function or NA. Function to break ties between max values. Ignored if mode is
not 'max'.}
}
\value{
a data frame with an extra variable containing a factor with the generated levels.
}
\description{
Constructs factor levels by assigning a value to combinations of columns. If mode is 'max', then
the columns will be treated as numeric and the associated factor level wiull be equal to the
name of the column contanining the maximum among the selected columns for each row. If mode is
'comb', factor levels are constructed by assigning a character value to the concatenation of
values contained in the selected columns, optionally reduced to flags s.t. x = x > 0 if
flags==TRUE.
}
