% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_util.R
\name{make_doc_set}
\alias{make_doc_set}
\title{Build document sets.}
\usage{
make_doc_set(obo, field, terms)
}
\arguments{
\item{obo}{An OBO interface object.}

\item{field}{A field over which to construct a document set.}

\item{terms}{A character vector containing terms to select documents for the DocSet.}
}
\value{
A DocSet instance that can be used to select observations from a corpus.
}
\description{
Document sets define the sample of corpus segments that will be considered for all counting
functions, particularly \code{\link{get_ cooccurrences}}.
}
\details{
If the given vector's length is greater than one, the filter will be constructed as the union of
all individual terms. If it is equal to one, it will be interpreted as a regular expression. If
its 0, the returned document set will have 0 documents.
}
