% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obo_helper.R
\name{make_obo}
\alias{make_obo}
\title{Create OBO interface object.}
\usage{
make_obo(home = getwd(), conf = make_obo_conf())
}
\arguments{
\item{conf}{An OBOConf instance, e.g. the value of \code{\link{make_obo_conf}}.}
}
\value{
An OBO index interface object.
}
\description{
The returned OBO instance object will take its parameters from the given conf object. If no conf
is given, one will be created with default parameters.
}
\details{
All index access functions require an OBO interface object as first parameter.
}
