% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obo_helper.R
\name{make_obo_conf}
\alias{make_obo_conf}
\title{Create a new OBO conf object}
\usage{
make_obo_conf()
}
\value{
An OBOConf instance.
}
\description{
The returned object may be used to modify corpus analysis parameters by using the native java
method 'set'. See examples.
}
\examples{
conf <- make_obo_conf()
conf@set( 'wPre', 10 ) # Set cooccurrence window leading offset to 10.
conf$set( 'wPos', 10 ) # Set cooccurrence window trailing offset to 10.
obo <- make_obo( conf ) # obtain an OBO index interface object.

}
