% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{random_spm}
\alias{random_spm}
\title{Create random sparse matrix.}
\usage{
random_spm(rows = 100, cols = 100, p = 0.1, v = function() 1)
}
\arguments{
\item{rows}{Integer vector of length 1. Number of rows.}

\item{cols}{Integer vector of length 1. Number of cols.}

\item{p}{Numeric vector of length 1. 0 < p < 1; fill rate.}

\item{v}{Numeric vector of length 1 or suplier function for values.}
}
\description{
Creates a random sparse matrix with the given dimension, fill rate and value function.
}
