% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_frequencies}
\alias{read_frequencies}
\title{Read term frequency data from DSV file.}
\usage{
read_frequencies(file, header = TRUE, sep = "@", lxcn = NULL)
}
\arguments{
\item{file}{Location of the frequencies file on disk.}

\item{header}{Logical indicating whether the given file contains a header as first row. Default
TRUE.}

\item{sep}{Character indicating the field separator value. Deafault '@'.}

\item{lxcn}{(optional) A lexicon data frame used for consistency checks.}
}
\value{
A data frame containing the list of terms as its row names, and as many columns as there
are corpus partitions containing each term's frequency counts for the respective corpus segment.
The returned data frame.
}
\description{
Reads term frequency data for all corpus partitions from a DSV file. Frequency data
is recorded on disk as a series of records containing each term's string form and an array of
frequency counts for all corpus partitions. Terms themselves are \emph{not} stored as data frame
columns, but are instead used as \emph{row names}.
}
