% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_lexicon}
\alias{read_lexicon}
\title{Read lexicon data from DSV file.}
\usage{
read_lexicon(file, header = TRUE, sep = "@")
}
\arguments{
\item{file}{Location of the lexicon file on disk.}

\item{header}{Logical indicating whether the given file contains a header as first row. Default
TRUE.}

\item{sep}{Character indicating the field separator value. Deafault '@'.}
}
\value{
A data frame containing the list of terms as its row names, a tf column for total term
frequencies and a df column for total document frequency for each term. The returned data frame
may contain additional columns for extra corpus-wide statistics depending on the corpus source
implementation.
}
\description{
Reads lexicon data from a DSV file. Lexicon data is recorded on disk as a series of
records containings each term's string form, its term total frequency and its total document
frequency. Additional columns may be used for additional corpus-wide term statistics. Terms
themselves are \emph{not} stored as data frame columns, but are instead used as row names.
Indices of terms in the vector produced by names( lexicon ) are used across all corpus datasets
as term keys.
}
