% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_pos_counts}
\alias{read_pos_counts}
\title{Read POS count data from DSV file.}
\usage{
read_pos_counts(file, header = TRUE, sep = "@", drop = TRUE)
}
\arguments{
\item{file}{Location of the POS count data file on disk.}

\item{header}{Logical indicating whether the given file contains a header as first row. Default
TRUE.}

\item{sep}{Character indicating the field separator value. Deafault '@'.}
}
\value{
A data frame containing the list of terms as its row names, and as many columns as there
are POS groups in the corresponding tagset, with each term's counts on each POS group.
}
\description{
Reads POS count data from a DSV file. POS count data is recorded on disk as a series
of records containings each term's string form, followed by one column for each major POS group
in the tagset used by the upstream corpus source for its NLP pipeline. Terms themselves are
\emph{not} stored as data frame columns, but are instead used as \emph{row names}.
}
