% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{str_normalize}
\alias{str_normalize}
\title{Normalize strings.}
\usage{
str_normalize(x, trim = TRUE, lower = TRUE, white = TRUE, nl = FALSE,
  punct = FALSE)
}
\arguments{
\item{x}{A string (i.e. a character vector of length 1).}

\item{trim}{Logical. Remove trailing and leading whitespace.}

\item{lower}{Logical. Reduce everything to lower case.}

\item{white}{Logical. Eliminate duplicate whitespace.}

\item{nl}{Loigcal. Replace newlines with spaces.}

\item{punct}{Logicel. Remove punctuation.}

\item{word}{Logical. Remove non-word characters.}
}
\value{
a reasonably normalized string.
}
\description{
This function offers a consistent procedure for string normalization combining several common
operations into one call.
}
\examples{
\\code{
x <- "some Ugly    dirty\\nstring!"
cat( x )
cat( normalizeString( x ) )
}
}
