% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vector_cosine}
\alias{vector_cosine}
\title{Cosine distances.}
\usage{
vector_cosine(m_, transpose = FALSE)
}
\arguments{
\item{m_}{A matrix of column (or row) vectors for which to compute distances.}

\item{transpose}{A logical vector indicating if distances should be computed across row vectors.
False by default.}
}
\value{
A square symmetric matrix of dimension equal to the column (or row) dimension of the
        given input matrix of cosine distances between the column (or row) vectors in the input
        matrix.
}
\description{
Compute the cosine distance between the given vectors. The cosine distance between vectors
i and j is equal to the dot product between them divided by the product of their norms:
\eqn{ ( i \dot j ) / [[i]]*[[j]] }.
}
\details{
The given matrix will be interpreted as an array of column vectors for which cosine distances
will be computed. If transpose == TRUE, the matrix will be interpreted as an array of row
vectors. In any case, the resulting matrix will be square, symmetric, and with dimensions
equal to the number of (column or row) vectors.
}
